♻️ refactor: item 타입 생성 및 그에 따른 코드 수정 #285
