+++ client/src/types/cart.type.ts
@@ -1,4 +1,4 @@
-import { Item } from './itemList.type';
+import { ItemShortcutData, CartItems, CartItemWithData } from './item.type';
 
 export interface CartData {
 	data: {
@@ -12,31 +12,8 @@ export interface CartData {
 	};
 }
 
-export interface CartItems {
-	data: CartItemWithData[];
-	pageInfo: null;
-}
-
-export interface CartItemWithData {
-	itemCartId: number;
-	quantity: number;
-	period: number;
-	buyNow: boolean;
-	subscription: boolean;
-	item: CartItem;
-	createdAt: Date;
-	updatedAt: Date;
-}
-
-export type CartItem = Omit<
-	Item,
-	'content' | 'starAvg' | 'reviewSize' | 'nutritionFacts' | 'discountPrice'
-> & {
-	disCountPrice: number;
-};
-
 export interface CartListProps {
 	data: CartItemWithData;
-	item: CartItem;
+	item: ItemShortcutData;
 	sub?: boolean;
 }
